+++ modulo4/servicos-backend/request.rest
@@ -1,0 +1,4 @@
+GET http://localhost:3003/address/90640220
+
+####
+POST http://localhost:3003/user/address/90640220
