+++ modulo4/servicos-backend/src/data/connection.ts
@@ -1,0 +1,24 @@
+import knex from 'knex'
+import dotenv from 'dotenv'
+import { Address } from "../types"
+
+
+
+dotenv.config();
+
+export const connection = knex({
+	client: "mysql",
+	connection: {
+    host: process.env.DB_HOST,
+    port: 3306,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: process.env.DB_NAME
+  }
+});
+
+export const insertAddress = async (address: Address) => {
+    await connection
+        .insert(address)
+        .into("aula50_address")
+}
