+++ modulo4/servicos-backend/src/index.ts
@@ -1,21 +1,7 @@
 import express, { Express } from "express";
-import knex from "knex";
 import cors from "cors";
-import dotenv from "dotenv";
 import { AddressInfo } from "net";
 
-dotenv.config();
-
-export const connection = knex({
-	client: "mysql",
-	connection: {
-    host: process.env.DB_HOST,
-    port: 3306,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASS,
-    database: process.env.DB_NAME
-  }
-});
 
 const app: Express = express();
 app.use(express.json());
