+++ modulo4/servicos-backend/src/types.ts
@@ -1,0 +1,9 @@
+export type Address = {
+    cep: string
+    logradouro: string
+	numero: string
+	complemento?: string,
+	bairro: string
+	cidade: string
+	estado: string
+}
